refactor: single source for app token
Import app token from single place for all
instances - examples, doc examples, doc live
editors.
