+++ doc/docs/intro/try.mdx
@@ -22,7 +22,7 @@ MapillaryJS is essentially an [npm package](https://www.npmjs.com/package/mapill
 For this guide, we will use the following tools.
 
 - You have [Node](https://nodejs.org) installed at `version >= v12.3.0`.
-- You have [Yarn](https://classic.yarnpkg.com) installed at `version >= v1.2.0`.
+- You have [Yarn](https://classic.yarnpkg.com) installed at `version >= v1.22.0`.
 
 To start using MapillaryJS with data from the [Mapillary](https://www.mapillary.com) platform, you need an app token. When [extending MapillaryJS](/docs/extend) to render your own data, no app token is needed.
 
@@ -191,7 +191,7 @@ function renderMapillary(props) {
 
   return (
     <ViewerComponent
-      appToken={'QjI1NnU0aG5FZFZISE56U3R5aWN4ZzpkYzg0NzE3MDA0YTRhZjlh'}
+      appToken={appToken}
       imageId={'ie9ktAVyhibDCD_V0m6apQ'}
       style={{width: '100%', height: '300px'}}
     />
