--- examples/umd.html
@@ -1,46 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <title>MapillaryJS UMD</title>
-    <link rel="icon" href="data:,">
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, user-scalable=no">
-
-    <link rel="stylesheet" href="../dist/mapillary.css" />
-
-    <script src="../dist/mapillary.js"></script>
-
-    <style>
-        body {
-            margin: 0;
-            padding: 0;
-        }
-
-        html,
-        body,
-        .viewer {
-            width: 100%;
-            height: 100%;
-        }
-    </style>
-</head>
-
-<body>
-    <script>
-        const { Viewer } = mapillary;
-
-        const container = document.createElement('div');
-        container.classList.add('viewer');
-        document.body.append(container);
-
-        const viewer = new Viewer({
-            apiClient: 'QjI1NnU0aG5FZFZISE56U3R5aWN4ZzpkYzg0NzE3MDA0YTRhZjlh',
-            component: { cover: false },
-            container,
-            imageId: 'zarcRdNFZwg3FkXNcsFeGw',
-        });
-    </script>
-</body>
-
-</html>
